Update nvidia-container.te
Running `sudo` inside a `nvidia_container_t` context failed with errors since the policy was missing the rules for sending audit messages. 
The other part why it failed was that the capability `CAP_AUDIT_WRITE` is not enabled by default. We need to run  the invocation with `--cap-add=AUDIT_WRITE`. 

With those two changes we have a clean `sudo mkdir DIR` inside of the container. 

```
 podman run  --cap-add=AUDIT_WRITE --security-opt label=type:nvidia_container_t --rm -it selinux sudo mktemp                                                                                                                                                              
/tmp/tmp.1XpA6b2LCV
```